requirements.txt
python-telegram-bot
toolz
algebraic-data-types
pydantic
tinycss2
